Got a bit side tracked form the ccache efficiency work when I found this optimization opportunity. It makes the OpenTitan model 8.5% faster when compiled without --trace (measured on a single threaded model), as V3Localize now finds an additional ~21000 variables that it can convert into locals which no longer need to live in (and be stored to) the model state. ~~(Sadly when compiled with tracing these opportunities go away as variables are read in the tracing routines.)~~ 

There are 3 patches and they should be kept separate as they are largely orthogonal. The first moves V3Descpoe after V3Localize, and teaches V3Localize how to localize variables of another module. Patch 2 is just infrastructure to be used in patch 3 initially but is generic. Patch 3 teaches V3Localize how to localize any variable that is always assigned before it's read within a function (previously it only did this if the variable was used only in a single function, but a lot of stuff was duplicaed into the settle loop, so there was a lot of lost opportuity).

Some numbers for OpenTitan (single threaded, without --trace):
```
Old Localize: (0a28fc8c)

      68,728.68 msec task-clock                #    1.000 CPUs utilized            ( +-  0.05% )
             75      context-switches          #    0.001 K/sec                    ( +- 31.50% )
              0      cpu-migrations            #    0.000 K/sec
          1,549      page-faults               #    0.023 K/sec                    ( +-  0.02% )
240,412,856,507      cycles                    #    3.498 GHz                      ( +-  0.05% )
232,031,502,038      instructions              #    0.97  insn per cycle           ( +-  0.00% )
 11,966,905,530      branches                  #  174.118 M/sec                    ( +-  0.00% )
    261,356,240      branch-misses             #    2.18% of all branches          ( +-  0.47% )

        68.7321 +- 0.0323 seconds time elapsed  ( +-  0.05% )


New Localize: (21917abf)

      63,407.49 msec task-clock                #    1.000 CPUs utilized            ( +-  0.04% )
             39      context-switches          #    0.001 K/sec                    ( +- 56.85% )
              0      cpu-migrations            #    0.000 K/sec
          1,434      page-faults               #    0.023 K/sec                    ( +-  0.01% )
221,799,502,259      cycles                    #    3.498 GHz                      ( +-  0.04% )
211,109,106,971      instructions              #    0.95  insn per cycle           ( +-  0.00% )
 11,898,040,540      branches                  #  187.644 M/sec                    ( +-  0.00% )
    264,866,360      branch-misses             #    2.23% of all branches          ( +-  0.78% )

        63.4091 +- 0.0265 seconds time elapsed  ( +-  0.04% )

```