@gezalore 2021-06-17
Actually, turns out even wiht tracing there are 19000 more locals in OpenTitan with this patch!