@wsnyder 2021-06-17
I assume you mean there are less locals with --trace with this patch?