@gezalore 2021-06-17
To clarify, on OpenTitan:
- Without this patch, V3Localize localizes about 1500 variables.
- With this patch, with --trace, it's about 20500. 
- With this patch, without --trace, it's about 22000.
Let's hope that stays that way after I fix the sanitizer failures..