@wsnyder 2021-06-17
Sounds like great results, wish had spotted that earlier!  Please also check SweRV performance.

1. Merging first change seems ok.
2. Like the concept, but perhaps we should apply it more globally to existing code, and as a separate pull?
3. Also seems reasonable.