@gezalore 2021-06-17
The sanitizer error seems a pre-existing genuine read past the end of a VlWide (trying to do a bit extract I think). It shows up now because the VlWide has been localized, while previously it just read past the end of it in the module struct.. Will need to figure out how to fix this.

> 2. Like the concept, but perhaps we should apply it more globally to existing code, and as a separate pull?

Agree it should be applied more globally where possible, but the last patch depends on this so I will keep it here for now, and revisit where the allocators can be used as a later patch/PR.