@gezalore 2021-06-17
> Please also check SweRV performance.

Somewhat more modes gain on EH1:

```
Old Localize:0a28fc8c

      18,299.03 msec task-clock                #    0.998 CPUs utilized            ( +-  0.03% )
             15      context-switches          #    0.001 K/sec                    ( +- 13.40% )
              0      cpu-migrations            #    0.000 K/sec
          1,398      page-faults               #    0.076 K/sec
 64,006,751,514      cycles                    #    3.498 GHz                      ( +-  0.03% )
 46,452,448,956      instructions              #    0.73  insn per cycle           ( +-  0.01% )
  4,547,912,648      branches                  #  248.533 M/sec                    ( +-  0.03% )
     58,756,429      branch-misses             #    1.29% of all branches          ( +-  0.39% )

       18.33155 +- 0.00519 seconds time elapsed  ( +-  0.03% )


New Localize: 21917abf

      18,138.79 msec task-clock                #    0.998 CPUs utilized            ( +-  0.04% )
             18      context-switches          #    0.001 K/sec                    ( +-  6.57% )
              0      cpu-migrations            #    0.000 K/sec
          1,398      page-faults               #    0.077 K/sec
 63,446,485,184      cycles                    #    3.498 GHz                      ( +-  0.04% )
 45,207,556,286      instructions              #    0.71  insn per cycle           ( +-  0.01% )
  4,427,565,512      branches                  #  244.094 M/sec                    ( +-  0.01% )
     58,920,165      branch-misses             #    1.33% of all branches          ( +-  0.36% )

       18.17047 +- 0.00724 seconds time elapsed  ( +-  0.04% )
```

On the other hand, OpenTitan builds about 20% faster on GCC with this patch (with no ccache, 2191 CPU seconds vs 2651), something that @rswarbrick might be happy to hear.