@gezalore 2021-06-17
> 2. Like the concept, but perhaps we should apply it more globally to existing code, and as a separate pull?

BTW I intend to rebase this and keep the 3 commits separate instead of squasing them. Sadly GitHub doesn't support dependent PRs..